// packages/blender-protocol/src/manifest-fields.generated.ts
// This file is auto-generated by scripts/generate_stage_scene_ops.py
// Do not edit manually - run 'npm run generate-stage-scene-ops' to update

import { Type } from "typebox";

export const GeneratedCameraManifestFields = {
	focusDistance: Type.Optional(Type.Number({ minimum: 0.0, maximum: 1000000.0 })),
};

export const GeneratedLightManifestFields = {
	cutoffDistance: Type.Optional(Type.Number({ minimum: 0.0, maximum: 1000000.0 })),
};